update 디렉토리 바꾸기 전에 저장
